requirements.txt
opencv-python==4.8.1.78
numpy==1.25.2
Pillow==10.1.0
pandas==2.1.2
pyttsx3==2.90
dlib==19.24.2
face-recognition==1.3.0
openpyxl==3.1.2

